web/requirements/project.txt
# project.txt is a pip requirements file which describes the distributions
# required by your project to run.

--requirement=base.txt

# Put project-specific requirements here.
# See http://pip-installer.org/requirement-format.html for more information.

markdown
django-filer
django-reversion
django-cms
cmsplugin-filer
cmsplugin-markdown
#django-tastypie
git+https://github.com/toastdriven/django-tastypie.git#egg=tastypie
requests
requests-cache
twitter-text-py
twython
pyfaceb
